chore: convert radio group to index based stories import
