+++ packages/react-components/react-radio/src/stories/RadioGroup/RadioGroupControlledValue.stories.tsx
@@ -2,28 +2,28 @@ import * as React from 'react';
 import { Label } from '@fluentui/react-label';
 import { tokens } from '@fluentui/react-theme';
 import { useId } from '@fluentui/react-utilities';
-import { Radio, RadioGroup } from '../index';
+import { Radio, RadioGroup } from '@fluentui/react-radio';
 
-export const Disabled = () => {
+export const ControlledValue = () => {
+  const [value, setValue] = React.useState('banana');
   const labelId = useId('label');
   return (
     <div style={{ display: 'grid', gridRowGap: tokens.spacingVerticalS }}>
-      <Label id={labelId} disabled>
-        Favorite Fruit
-      </Label>
-      <RadioGroup defaultValue="apple" disabled aria-labelledby={labelId}>
+      <Label id={labelId}>Favorite Fruit</Label>
+      <RadioGroup value={value} onChange={(_, data) => setValue(data.value)} aria-labelledby={labelId}>
         <Radio value="apple" label="Apple" />
         <Radio value="pear" label="Pear" />
         <Radio value="banana" label="Banana" />
         <Radio value="orange" label="Orange" />
       </RadioGroup>
+      <div>Current value: {value}</div>
     </div>
   );
 };
-Disabled.parameters = {
+ControlledValue.parameters = {
   docs: {
     description: {
-      story: 'RadioGroup can be disabled, which disables all Radio items inside.',
+      story: 'The selected radio item can be controlled using the `value` and `onChange` props.',
     },
   },
 };
