+++ packages/react-components/react-radio/src/stories/RadioGroup/RadioGroupHorizontal.stories.tsx
@@ -2,26 +2,27 @@ import * as React from 'react';
 import { Label } from '@fluentui/react-label';
 import { tokens } from '@fluentui/react-theme';
 import { useId } from '@fluentui/react-utilities';
-import { Radio, RadioGroup } from '../index';
+import { Radio, RadioGroup } from '@fluentui/react-radio';
 
-export const DisabledItem = () => {
+export const Horizontal = () => {
   const labelId = useId('label');
   return (
     <div style={{ display: 'grid', gridRowGap: tokens.spacingVerticalS }}>
       <Label id={labelId}>Favorite Fruit</Label>
-      <RadioGroup defaultValue="apple" aria-labelledby={labelId}>
+      <RadioGroup layout="horizontal" aria-labelledby={labelId}>
         <Radio value="apple" label="Apple" />
         <Radio value="pear" label="Pear" />
-        <Radio value="banana" label="Banana" disabled />
+        <Radio value="banana" label="Banana" />
         <Radio value="orange" label="Orange" />
       </RadioGroup>
     </div>
   );
 };
-DisabledItem.parameters = {
+Horizontal.storyName = 'Layout: horizontal';
+Horizontal.parameters = {
   docs: {
     description: {
-      story: 'Radio items can be disabled individually.',
+      story: 'The `horizontal` layout places each radio item in a row, with labels after the radio indicator.',
     },
   },
 };
