+++ packages/react-components/react-radio/src/stories/RadioGroup/RadioGroupUncontrolledValue.stories.tsx
@@ -2,15 +2,14 @@ import * as React from 'react';
 import { Label } from '@fluentui/react-label';
 import { tokens } from '@fluentui/react-theme';
 import { useId } from '@fluentui/react-utilities';
-import { Radio, RadioGroup } from '../index';
+import { Radio, RadioGroup } from '@fluentui/react-radio';
 
-export const HorizontalStacked = () => {
+export const UncontrolledValue = () => {
   const labelId = useId('label');
-
   return (
     <div style={{ display: 'grid', gridRowGap: tokens.spacingVerticalS }}>
       <Label id={labelId}>Favorite Fruit</Label>
-      <RadioGroup layout="horizontal-stacked" aria-labelledby={labelId}>
+      <RadioGroup defaultValue="pear" aria-labelledby={labelId}>
         <Radio value="apple" label="Apple" />
         <Radio value="pear" label="Pear" />
         <Radio value="banana" label="Banana" />
@@ -19,11 +18,13 @@ export const HorizontalStacked = () => {
     </div>
   );
 };
-HorizontalStacked.storyName = 'Layout: horizontal-stacked';
-HorizontalStacked.parameters = {
+UncontrolledValue.parameters = {
   docs: {
     description: {
-      story: 'The `horizontal-stacked` layout places each radio item in a row, with labels below the radio indicator.',
+      story:
+        'The initially selected item can be set by setting the `defaultValue` of RadioGroup. ' +
+        'Alternatively, one Radio item can have `defaultChecked` set. ' +
+        'Both methods have the same effect, but only one should be used in a given RadioGroup.',
     },
   },
 };
